--- packages/notary-service-client/.depcheckrc
@@ -1,1 +1,0 @@
-ignores: []
